requirements.txt
python-telegram-bot[httpx] ~= 20.3
yt-dlp
python-dotenv